pyproject.toml
[project]
name = "OpenAnonymizer"
version = "0.1.0"
description = "Dutch-based Anonimyzer for PII data in pdfs."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
dependencies = [
    "fastapi[all]>=0.115.12",
    "presidio-analyzer>=2.2.358",
    "transformers>=4.51.3",
    "nl_core_news_lg",
    "torch>=2.7.0",
    "presidio-anonymizer>=2.2.358",
    "pikepdf>=9.9.0",
    "python-dotenv>=1.1.0",
    "pycryptodome>=3.23.0",
    "sqlalchemy>=2.0.41",
    "pymupdf>=1.26.3",
    "PyYAML>=6.0.2",
]
authors = [
    { name = "Mark Westerweel", email = "mark.westerweel@conduction.nl" },
    { name = "Razo van Berkel", email = "razo.van.berkel@centric.eu" },
]

[tool.uv.sources]
nl-core-news-lg = { url = "https://github.com/explosion/spacy-models/releases/download/nl_core_news_lg-3.8.0/nl_core_news_lg-3.8.0-py3-none-any.whl"}

[tool.ruff]
line-length = 88
target-version="py312"
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
extend-select = ["D", "I"]
ignore = ["D1"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
disallow_untyped_defs = true
disallow_any_unimported = false
implicit_optional = false
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
python_version = "3.12"
exclude = ["tests/"]
follow_untyped_imports = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[dependency-groups]
flair = [
    "flair>=0.15.1",
]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-json-report>=1.5.0",
]
